refactor : 모달창 코드 수정
