+++ frontend/src/components/Modal/AfterLogin.tsx
@@ -1,12 +1,18 @@
-import React from 'react';
+import React, {useState} from 'react';
 import BookCover from '../bookshape/BookCover';
 import Modals from './Modal'
+import OpenModal from './OpenModal';
 
 function AfterLogin() {
+  const [selected, setSelected] = useState<string>('images/mainLogo.png');
+  const [isOpen, setIsOpen] = useState<boolean>(false);
   return (
-    <BookCover>
-      <Modals />
-    </BookCover>
+    <>
+      <OpenModal isOpen={isOpen} setIsOpen={setIsOpen} setSelected={setSelected} />
+      <BookCover>
+        <Modals setIsOpen={setIsOpen} selected={selected} setSelected={setSelected} />
+      </BookCover>
+    </>
   );
 }
